scsynth: use boost::program_options
fixes #3117

this gets rid of the old fragile option parsing in favor of
boost::program_options. i tried to keep everything as consistent as
possible, but the result was not as clean as i was hoping.

a difficulty is that WorldOptions is in the public API, so the
names and types of arguments cannot be changed to help integrate with
boost::program_options. probably the best thing to do would be use
server_args in supernova for both, and have a function that converts
directly from server_args to WorldOptions

some more notable changes:

added long options (#4028) identical to supernova's

udp & tcp port defaults changed from -1 to 0
- impact low, not good practice to begin with (0-1023 are system ports)

-U can be specified multiple times, accumulating each time
- low-impact, previous invocations that used it multiple times and assumed the last option would be the only one to take effect will break

now an error if -L passed but memory locking not available
- this is good, because users should know if they are not getting the options they request